# Optimize slicing graph generation